move helpers to lib folder
