+++ app/modules/transactions/lib/helpers.js
